removed all the methods in FishBot that should not be used. Added tests <testing, feature
