+++ src/main/java/lasersharks/enemies/FishImage.java
@@ -1,0 +1,8 @@
+package lasersharks.enemies;
+
+public interface FishImage {
+  
+  public double getWidthScale();
+  
+  public String getImageResource();
+}
